industry_sentiment/backend/requirements.txt
akshare
pandas
psycopg2-binary
SQLAlchemy
requests
python-dateutil
